requirements.txt
torch
diffusers
transformers
accelerate
gradio
safetensors
pillow
numpy<2
invisible_watermark
huggingface_hub[hf_transfer]
sentencepiece
opencv-python==4.5.5.64
gguf
tensorflow